requirements.txt
selenium==4.1.3
pytest==7.1.2
chromedriver-binary==102.0.5005.61.0
allure-pytest==2.9.45